pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "bands-inspect"
version = "0.3.1"
description = "Utilities for creating, comparing and plotting bandstructures of materials."
authors = ["Dominik Gresch <mail@greschd.ch>"]
license = "Apache 2.0"
readme = "README.md"
documentation = "https://bands-inspect.greschd.ch"
repository = "https://github.com/Z2PackDev/bands_inspect"
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Operating System :: Unix",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Physics",
    "Development Status :: 4 - Beta",
]
keywords = [
    "condensed matter",
    "solid-state",
    "physics",
    "materials science",
]
packages = [
    { include = "bands_inspect", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
numpy = "^1.20"
scipy = "^1.7"
click = "^8.1"
fsc-export = "^1.2"
decorator = "^5.1"
matplotlib = "^3"
h5py = "^3.8"
fsc-hdf5-io = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
sphinx = "^6.1.3"
sphinx-rtd-theme = "^1.2.0"
sphinx-click = "^4.4.0"
ipython = "^8.11.0"
tbmodels = "^1.4.3"
pre-commit = "^3.1.1"
pylint = "^2.17.0"
mypy = "^1.1.1"
isort = "^5.12.0"
ruamel-yaml = "^0.17.21"


[tool.poetry.scripts]
bands-inspect = "bands_inspect._cli:cli"

[tool.poetry.plugins."fsc.hdf5_io.load"]
bands_inspect = "bands_inspect"
